requirements.txt
azure-identity
openai>=1.108.1
python-dotenv
langchain-text-splitters
